relocate reewc codes
