+++ sevenn/train/reewc/__init__.py
@@ -1,0 +1,16 @@
+from .loss import EWCLoss, append_ewc_loss
+from .rehearsal import (
+    build_memory_loader,
+    reewc_dataset_keys,
+    validate_reewc_config,
+)
+from .trainer import ReewcTrainer
+
+__all__ = [
+    'EWCLoss',
+    'append_ewc_loss',
+    'ReewcTrainer',
+    'build_memory_loader',
+    'reewc_dataset_keys',
+    'validate_reewc_config',
+]
